requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dotenv>=1.0.0
schedule>=1.2.0
pydantic>=2.0.0
httpx>=0.25.0
aiohttp>=3.8.0
flask>=2.3.0
flask-cors>=4.0.0
waitress>=2.1.0
